requirements.txt
pylint
prettytable
matplotlib.pyplot